Volta ao tempo e recuperacao de arquivo

+++ PastaFontes/Produto.java.java
@@ -1,21 +1,19 @@
-//Programa de aplicacao que administra o cadastro de produtos
-//Inclui e consulta dados de produto
 
 public class Produto {
 
-	private int peso;
+	private double peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(int peso, double altura) {
+	public Produto(double peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public int getPeso() {
+	public double getPeso() {
 		return peso;
 	}
